agents/customer-support/requirements.txt
# Customer Support Agent Runtime Dependencies
# Minimal set - transitive deps are resolved automatically

# Bedrock AgentCore SDK (provides runtime execution)
bedrock-agentcore>=1.0.0

# Strands framework (provides Agent, BedrockModel, tools, hooks)
strands-agents>=0.1.0

# MCP protocol for Gateway tools
mcp>=1.0.0

# Note: Workspace packages (agentcore-common, agentcore-tools) are installed
# from local sources during Docker build, not from PyPI
